requirements.txt
mss>=9.0.1
Pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.9.0
pyperclip>=1.8.2